update tests to be able to handle ZM output
